style: :lipstick: 스타일 코드 위치 변경
은행 리스트(셀렉트 박스를 위한) 조회 관련 코드는 Store로 변경
